In this iteration we renamed frontend2 to frontend in preparation for production deploy
