+++ frontend/app/components/PhotoGrid.vue
