+++ frontend/app/composables/useAuth.js
