+++ frontend/app/composables/useSingleImage.js
