+++ frontend/app/middleware/auth.global.js
