+++ frontend/app/pages/album/[albumid]/image/[id].vue
