+++ frontend/app/pages/album/[id].vue
