+++ frontend/app/pages/logout.vue
