--- frontend/store/alert.js
@@ -1,16 +1,0 @@
-export const state = () => ({
-  alertData: null
-})
-
-export const mutations = {
-  insertAlert(state, obj) {
-    // obj contains alertType, alertMessage
-    if (!obj.alertType) {
-      obj.alertType = 'error'
-    }
-    state.alertData = obj
-  },
-  clearAlert(state) {
-    state.alertData = null
-  }
-}
