--- frontend2/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "nuxt-app",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "build": "nuxt build",
-    "dev": "nuxt dev",
-    "generate": "nuxt generate",
-    "preview": "nuxt preview",
-    "postinstall": "nuxt prepare"
-  },
-  "dependencies": {
-    "nuxt": "^4.0.2",
-    "@mdi/font": "^7.4.47",
-    "@vite-pwa/nuxt": "^1.0.0",
-    "sass": "^1.89.0",
-    "vite-plugin-vuetify": "^2.1.1",
-    "vuetify": "^3.9.3"
-  }
-}
